Add files via upload
In this project, I attempted to diagnose with Machine Learning, if the given person is having a heart disease or not. Upon further, importing of the data, I used plot to analyze the data. After that, to deal with categorical features, I created dummy variables and for numerical features, I applied scaling. I applied six Machine Learning algorithms: Logistic Regression 0.78, K Neighbors Classifier 0.73, Support Vector Classifier 0.80, Gradient boosting Classifier 0.80, Decision Tree 0.72, Random Forest Classifier 0.83. I modified settings of each of the models to enhance their performances. Therefore, from the above said study the implementation of Random Forest Classifier yielded the highest percentage of 83%, which revealed the best performance.

When the user run GUI and has entered his medical data the GUI of the application will show the prediction on the Random Forest Classifier classification models. The output contains the predicted class (No heart disease class is 0 while the class with heart disease is 1) and the possibility of that class.